IDE versioning logic for layout strategy/policy
